requirements.txt
fastapi
uvicorn
pillow
pydantic
lumaai
google-generativeai
pyyaml
